Added stripping and inner_xml functions with doctests.
The stripped_inner_xml test does not pass, however, as it demonstrates
clearly that the result we get back is not the pretty one I'd desire.

What I'd like to get from:

  <div>
     This is some <i><b>really</b>
  silly</i> text!</div>

is:

  <div>This is some <i><b>really</b> silly</i> text!</div>

However, what we currently get is:

  <div>This is some<i><b>really</b>silly</i>text!</div>

This isn't correct... we need to respect spaces before and after
certain elements, I think... I need to figure out how to do that
cleanly.

Otherwise else, we're going to end up with some really nasty-looking
templates.